Implement search along all interesting files/folders
We are in a project state the we need to look for a fish if it can be
found in a certain jar, if it has already be scanned, if it is in the
list of full heads to be scanned and other things.
Implementing some kind of search function over different locations of
interest.