Keeping track of FullHeadList.txt
Mikki and Kassandra compiled a list of fishes which need a full head
scan.
The latter part of the notebook goes through that list and helps us
finding the few fishes left to scan.
This should minimize fish searching time, as this has now become an annoying
part of the time spent on the project :)