Repeat after me: "never re-use variable names!"
Direction is for the fishes, plane is for the images!
We used to use direction for both, which lead to weird issues.

Now we are correctly extracting the gray values for each MIP for each
plane.
Using the signal library of scipy we extract the peaks and the width,
which make it fairly easy to extract the otolith region.